// libs/shared/ui/src/lib/clientes/cliente-busqueda.dto.ts
export interface ApiClienteDatosPersonalesRequest {
  readonly documento: string;
  readonly usuario: string;
  readonly aplicacion: string;
  readonly consultarProspect?: string;
  readonly idioma?: string;
}

export interface ApiCliente {
  readonly clienId?: number;
  readonly documento?: string;
  readonly tipoPersona?: string;
  readonly nombre?: string;
  readonly apell1?: string;
  readonly apell2?: string;
  readonly razonSocial?: string;
  readonly fecNac?: string;
  readonly indSexo?: string;
}

export interface ApiDatosBusquedaCliente {
  readonly clientes?: readonly ApiCliente[];
  readonly prospectoClientes?: readonly ApiCliente[];
}